+++ NEWS.md
@@ -3,6 +3,7 @@
 ## hydraulics (development version)
 
   - Added functionality to solve for absolute roughness
+  - Added manningc function to include partially filled pipe solutions
 
 ## hydraulics 0.1.0